Add type checking, refactor locations of setting enum
