+++ .github/workflows/test.yml
@@ -30,4 +30,5 @@ jobs:
       - run: yarn format
       - run: yarn lint
       - run: yarn test
+      - run: yarn typecheck
       - run: yarn build
