+++ src/features/rap/extra/settings/settingEnums.ts
@@ -1,0 +1,42 @@
+export enum AltitudeType {
+  AGL = "AGL",
+  MSL = "MSL",
+}
+
+export enum AltitudeLevels {
+  Default = "Default",
+  Normalized = "Normalized",
+}
+
+// CSS scroll-snap-stop
+export enum OnOff {
+  On = "On",
+  Off = "Off",
+}
+
+export enum SpeedUnit {
+  KPH = "km/h",
+  MPH = "mph",
+  Knots = "knots",
+  mps = "m/s",
+}
+
+export enum HeightUnit {
+  Feet = "ft",
+  Meters = "m",
+}
+
+export enum TemperatureUnit {
+  Celsius = "°C",
+  Fahrenheit = "°F",
+}
+
+export enum DistanceUnit {
+  Miles = "miles",
+  Kilometers = "km",
+}
+
+export enum TimeFormat {
+  TwentyFour = "24-hour",
+  Twelve = "12-hour",
+}
